requirements.txt
# Core dependencies
networkx>=3.1
rdflib>=7.0.0
plantuml-markdown>=3.9.2

# LLM Integration
openai>=1.0.0
anthropic>=0.7.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1

# Code Quality
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0

# Utilities
pyyaml>=6.0.1
python-dotenv>=1.0.0
click>=8.1.7
pydantic>=2.4.0
